PoC-2/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
sqlalchemy==2.0.36
alembic==1.14.0
aiosqlite==0.20.0
openai==1.59.4
python-multipart==0.0.12
jinja2==3.1.4
python-dotenv==1.0.1
twilio==9.6.3
sendgrid==6.12.2
httpx==0.28.1
